+++ pyinfra/api/connectors/__init__.py
@@ -1,6 +1,4 @@
-import six
-
-from . import ansible, docker, local, mech, ssh, vagrant
+from . import ansible, docker, local, mech, ssh, vagrant, winrm
 
 
 # Connectors that handle execution of pyinfra operations
@@ -8,18 +6,16 @@ EXECUTION_CONNECTORS = {  # pragma: no cover
     'docker': docker,
     'local': local,
     'ssh': ssh,
+    'winrm': winrm,
 }
 
 # Connectors that handle generation of inventories
-INVENTORY_CONNECTORS = {  # pragma: no cover
-    'ssh': ssh,
-    'docker': docker,
-    'mech': mech,
-    'vagrant': vagrant,
+ALL_CONNECTORS = {  # pragma: no cover
     'ansible': ansible,
+    'docker': docker,
+    'local': local,
+    'mech': mech,
+    'ssh': ssh,
+    'vagrant': vagrant,
+    'winrm': winrm,
 }
-
-ALL_CONNECTORS = (  # pragma: no cover
-    list(six.iterkeys(EXECUTION_CONNECTORS))
-    + list(six.iterkeys(INVENTORY_CONNECTORS))
-)