+++ pyinfra/api/state.py
@@ -86,8 +86,11 @@ class State(object):
 
     # Flags for printing
     print_output = False  # print output from the actual deploy (-v)
+    print_input = False
+
     print_fact_info = False  # log fact gathering as INFO > DEBUG (-v)
     print_fact_output = False  # print output from facts (-vv)
+    print_fact_input = False
 
     # Used in CLI
     deploy_dir = None  # base directory for locating files/templates/etc