+++ pyinfra_cli/legacy.py
@@ -43,6 +43,10 @@ Config options:
     -u --user USER       SSH user.
     --key FILE           SSH private key.
     --key-password PASS  SSH key password.
+    --winrm_password PWD WINRM password (bad practice).
+    --winrm_username USR WINRM username.
+    --shell_executable   Default shell to use (ex: 'sh' (default), 'ps', 'cmd')
+    --winrm_port PORT    WINRM port.
     --sudo               Use sudo.
     --sudo-user USER     Which user to sudo to.
     --su-user USER
@@ -256,6 +260,9 @@ def setup_arguments(arguments):
 
         # Config options
         'user': arguments['--user'],
+        'winrm_username': arguments['--winrm_username'],
+        'winrm_password': arguments['--winrm_password'],
+        'winrm_port': arguments['--winrm_port'],
         'key': arguments['--key'],
         'key_password': arguments['--key-password'],
         'password': arguments['--password'],